feat: 페이지당 row 개수 상태 관리
